--- demo1.py
@@ -1,24 +1,0 @@
-# from WindPy import w
-#
-# w.start()  # 默认命令超时时间为120秒，如需设置超时时间可以加入waitTime参数，例如waitTime=60,即设置命令超时时间为60秒
-#
-# w.isconnected()  # 判断WindPy是否已经登录成功
-# what = w.wsd('600030.SH', 'CLOSE', '2016-01-01', '2016-01-05', '')
-# print(what)
-# 我随便在开发分支里写点东西
-
-
-# import datetime,time
-# res = datetime
-# res_time = time.strftime("%Y-%m-%d %H:%M:%S", time.localtime())
-# res_type = datetime.datetime.strptime(res_time,"%Y-%m-%d %H:%M:%S")
-# print(res_time)
-# print(type(res_time))
-# print(res_type)
-# print(type(res_type))
-
-for i in range(1,10):
-    print(i)
-STR = "GSGAGSGS6"
-S = STR[:-1]
-print(S)
