+++ testException/testRead/read_Excel_2.py
@@ -7,7 +7,8 @@ from testException.utils.MysqlProxy import MysqlProxy
 base_dir = r'D:\Users\Administrator\AppData\Local\Programs\PycharmWorkSpace\Test_Dev\testException\origin'
 #file_name = '20220824_(XY6305)久铭专享23号私募证券投资基金_证券投资基金估值表.xls'
 #file_name = 'SW8251_久铭6号私募证券投资基金_2022-08-26估值表.xls'
-file_name = '估值表日报-STV732-静久康铭2号私募证券投资基金-4-20220901.xlsx'
+file_name = '估值表日报-STV732-静久康铭2号私募证券投资基金-4-20220901.xls'
+
 #datestr,pro_code,pro_name = re.search(r'(\d+)_\W([A-Za-z0-9]+)\W(\w+)私募', file_name).groups()
 #2022-08-29_SL0795_久铭稳健22号私募证券投资基金估值表.xlsx [国信]
 #20220826_(XY6307)久铭专享21号私募证券投资基金_证券投资基金估值表.xls
@@ -27,8 +28,8 @@ data_frame = pd.read_excel(os.path.join(base_dir,file_name))
 
 
 PRODUCT_CODE_NAME_MAP = {'SN0910':'久铭1号'}
-product_id, product, date_str = re.search(r'([A-Za-z0-9]+)(\w+)私募[^\d]+(\d+)', file_name).groups()
-product_code, product_name = product_id, PRODUCT_CODE_NAME_MAP[product_id]
+# product_id, product, date_str = re.search(r'([A-Za-z0-9]+)(\w+)私募[^\d]+(\d+)', file_name).groups()
+# product_code, product_name = product_id, PRODUCT_CODE_NAME_MAP[product_id]
 
 
 resss = data_frame.index
