+++ testException/testZZ/TestZhengZe.py
@@ -4,5 +4,5 @@ import re
 #respons = re.findall('\d+','30天多少个1字')#返回数字字符串的数组
 mo = r'[\u4e00-\u9fa5]*'  #=============>>匹配的是字符串首位出现的汉字字符串
 mo2= r'[\u4E00-\u9FA5]+'   #=============>>匹配整个字符串中第一次出现的汉字字符串
-res = re.search(mo2, 'fsdff汉字hfhe中国')
+res = re.search(mo2, '汉字hfhe中国').group(0)
 print(res)  # <re.Match object; span=(0, 2), match='我是'>
